+++ src/main/java/me/gurkz/superdupermod/mixin/LeashableVillagerMixin.java
@@ -8,7 +8,7 @@
 
 package me.gurkz.superdupermod.mixin;
 
-import net.minecraft.world.entity.npc.AbstractVillager;
+import net.minecraft.world.entity.npc.villager.AbstractVillager;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;